pyproject.toml
[project]
name = "ticket-retrieval-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "jq>=1.11.0",
    "langchain>=1.2.11",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-openai>=1.1.11",
    "python-dotenv>=1.0.0",
    "streamlit>=1.40.0",
    "uvicorn>=0.41.0",
]
